requirements.txt
python-telegram-bot
python-gitlab
jira
python-dotenv 